mic_serv_users/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.15
botocore==1.31.15
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.1
frozenlist==1.4.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
PyJWT==2.8.0
pymongo==4.4.1
pymongo-auth-aws==1.1.0
python-dateutil==2.8.2
pytz==2023.3
PyJWT==2.8.0
phonenumbers==8.13.20
passlib==1.7.4
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
twilio==8.5.0
urllib3==1.26.16
yarl==1.9.2
